refactor(model-selector): extract fuzzy search to shared modelSearch util
Move the fuzzy match helpers out of useModelSelectorData into
src/renderer/utils/modelSearch.ts so other surfaces can reuse the
same scoring logic. useModelSelectorData now wraps the shared
scorer via a small getModelSearchScore helper; behavior is
unchanged for the model selector.
